Constructive feedback from grader
Ralph, Good job on the homework assignment. You've demonstrated your best tactics for network vulnerability assessment and mitigation. In phase 1’s deliverable, you've captured and included your fping commands, the results of each command, and a listing of IPs returning echo replies. You’ve done an excellent job with phase 2 as well. With the hosts found in phase 1, you've determined which ports are open. Good work; the summary depicts the above! Well done on phase 3. With your findings from phase 2, you've determined whether or not you can connect and log in to the pingable server. Your summary is comprehensive, and includes an explanation for the misdirection, the query command, and the corresponding name. Phase 4 is an awesome demonstration of your sleuthing and Wireshark skills. Your findings include the name of the capture file, analysis of the HTTP traffic and hacker’s message, and the rogue MAC address. In each phase, you’ve named the correct OSI model in question, and you've offered fantastic high-level mitigation recommendations. Nice work!
 BLS
Central Grader , Apr 7, 2023 at 9:49am